requirements.txt
prettytable
colorama
python-telegram-bot
requests
schedule
python-dotenv